requirements.txt
tensorflow>=2.8.0
opencv-python>=4.5.0
numpy>=1.19.0
tqdm>=4.62.0
face-recognition>=1.3.0
termcolor>=1.1.0
ascii-magic>=2.1.0
matplotlib>=3.4.0
Pillow>=8.0.0 